fix: swagger-ui & update-method
After hapi-swagger 10.x update, it uses swagger-ui 3.x under the hood and jsonEditor was removed. Before removing this config options the whole /documentation page did not work.
When testing all the API methods, I noticed that update crashes the whole service because nedb threw an error. I fixed the actual issue now, but the service should not crash because of these. That needs to be looked at separately.